+++ rezasm-source/rezasm-core/src/instructions/implementation/mod.rs
@@ -1,10 +1,10 @@
-pub mod arithmetic_instructions;
-pub mod branch_instructions;
-pub mod comparison_instructions;
-pub mod float_arithmetic_instructions;
-pub mod function_instructions;
-pub mod memory_instructions;
-pub mod terminal_output_instructions;
+mod arithmetic_instructions;
+mod branch_instructions;
+mod comparison_instructions;
+mod float_arithmetic_instructions;
+mod function_instructions;
+mod memory_instructions;
+mod terminal_output_instructions;
 
 pub fn register_instructions() {
     arithmetic_instructions::register_instructions();
@@ -14,4 +14,142 @@ pub fn register_instructions() {
     function_instructions::register_instructions();
     memory_instructions::register_instructions();
     terminal_output_instructions::register_instructions();
+}
+
+#[cfg(test)]
+mod tests {
+    use std::fs;
+    use std::io::Write;
+    use std::ops::Deref;
+
+    use arithmetic_instructions::ADD;
+    use crate::parser::lexer::parse_lines;
+    use crate::test_utils::{workspace_root, TestWriter};
+
+    use crate::util::io::RezasmFileReader;
+    use crate::{
+        parser::line::Line,
+        simulation::{registry, simulator::Simulator},
+    };
+
+    use super::*;
+
+    // Moved from Trevor's implementation in tests/core.rs
+    #[test]
+    pub fn test_macro() {
+        register_instructions();
+        let mut simulator: Simulator = Simulator::new();
+
+        let line: Line = Line::new(
+            &String::from("add"),
+            ["$T0".to_string(), "$T0".to_string(), "121".to_string()].to_vec(),
+            simulator.get_word_size(),
+        )
+        .unwrap();
+
+        let args = match line {
+            Line::Instruction(_, args) => args,
+            _ => Vec::new(),
+        };
+
+        let foo = &ADD;
+
+        match foo.call_function(&mut simulator, &args) {
+            Ok(_) => {
+                assert_eq!(
+                    simulator
+                        .get_registers()
+                        .get_register(&registry::T0.to_string())
+                        .unwrap()
+                        .get_data()
+                        .int_value(),
+                    121
+                );
+            }
+            Err(_) => {
+                assert!(false);
+            }
+        }
+    }
+
+    // Moved from Trevor's test in tests/core.rs
+    #[test]
+    pub fn test_print_instructions() {
+        register_instructions();
+        let writer = Box::new(TestWriter::new());
+        let mut simulator: Simulator = Simulator::new_writer(writer);
+        let program = "
+        move $s2 \"Print Instructions Work!\\n\"
+        add $s1 '\\n' 0
+        add $t0 1 2
+        printi $t0
+        printc $s1
+        addf $t1 1.5 0
+        printf $t1
+        printc $s1
+        prints $s2";
+        let lines = parse_lines(&program.to_string(), simulator.get_word_size()).unwrap();
+        simulator.add_lines(lines, "".to_string()).unwrap();
+        while !simulator.is_done() {
+            simulator.run_line_from_pc().unwrap();
+        }
+        let writer = simulator.get_writer();
+        let output = writer
+            .deref()
+            .as_any()
+            .downcast_ref::<TestWriter>()
+            .unwrap()
+            .get_data();
+        assert_eq!(output.as_str(), "3\n1.5\nPrint Instructions Work!\n");
+    }
+
+    // Moved from Trevor's test in tests/core.rs
+    #[test]
+    pub fn test_io() {
+        register_instructions();
+
+        // Read into rezasm file
+        let file_path = workspace_root().join("example").join("arithmetic_fib.ez");
+        let mut rezasmfile = RezasmFileReader::new(file_path.clone()).expect("failed to read file");
+
+        // Reads with fs
+        let reg_bytes = fs::read(file_path).unwrap();
+        let reg_read: Vec<String> = String::from_utf8(reg_bytes.clone())
+            .unwrap()
+            .lines()
+            .map(|line| line.to_string())
+            .collect();
+
+        // Compare
+        assert_eq!(rezasmfile.lines().expect("failed to get lines"), reg_read);
+
+        assert_eq!(32, rezasmfile.seek_absolute_byte(1).unwrap());
+        assert_eq!(101, rezasmfile.seek_absolute_byte(3).unwrap());
+        assert_eq!(32, rezasmfile.seek_relative_byte(-2).unwrap());
+        assert_eq!(35, rezasmfile.peek_absolute_byte(0).unwrap());
+
+        rezasmfile.seek_start();
+
+        let mut bytes = vec![];
+        while let Some(byte) = rezasmfile.next() {
+            bytes.push(byte);
+        }
+
+        // Compare the bytes.
+        assert_eq!(reg_bytes, bytes);
+
+        // Cursor should now be at the end. Attempt to peek 0.
+        assert_eq!(35, rezasmfile.peek_absolute_byte(0).unwrap());
+
+        // Check validity (should be invalid).
+        assert_eq!(false, rezasmfile.is_cursor_valid());
+
+        // Convert to a writer.
+        let file_path = workspace_root().join("example").join("arith_modified.ez");
+        let path: String = file_path.to_string_lossy().into();
+        let mut writer = rezasmfile.into_writer(Some(path)).unwrap();
+        let write_str = "\n# This comment was generated by io tests!";
+        writer.extend_from_slice(write_str.as_bytes());
+        writer.flush().unwrap();
+    }
 }
