+++ rezasm-source/rezasm-core/src/simulation/memory.rs
@@ -190,3 +190,23 @@ impl Memory {
         Ok(out)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // Moved from Trevor's test in tests/core.rs
+    #[test]
+    fn test_memory() {
+        let mut memory: Memory = Memory::new();
+        let data = RawData::from_int(100, &DEFAULT_WORD_SIZE);
+        memory.write(memory.current_heap_pointer(), &data).unwrap();
+        assert_eq!(
+            memory
+                .read(memory.current_heap_pointer())
+                .unwrap()
+                .int_value(),
+            100
+        );
+    }
+}
