+++ rezasm-source/rezasm-core/src/test_utils/writer.rs
@@ -1,5 +1,5 @@
-use rezasm_core::simulation::writer::Writer;
-use rezasm_core::util::as_any::AsAny;
+use crate::simulation::writer::Writer;
+use crate::util::as_any::AsAny;
 use std::any::Any;
 use std::fmt::Debug;
 use std::io::Write;
